requirements.txt
imaplib
email
openai
pandas
datetime
python-dotenv
ast
os
json
bs4
streamlit